feat(site): dry up AI Bridge setup alert
