+++ site/src/pages/AIBridgePage/ListSessionsPage/ListSessionsPageView.tsx
@@ -1,8 +1,6 @@
 import { InfoIcon } from "lucide-react";
 import type { ComponentProps, FC, PropsWithChildren } from "react";
 import type { AIBridgeSession } from "#/api/typesGenerated";
-import { Alert, AlertDescription, AlertTitle } from "#/components/Alert/Alert";
-import { Link } from "#/components/Link/Link";
 import {
 	PaginationContainer,
 	type PaginationResult,
@@ -23,8 +21,8 @@ import {
 	TooltipProvider,
 	TooltipTrigger,
 } from "#/components/Tooltip/Tooltip";
-import { docs } from "#/utils/docs";
 import { DATE_FORMAT, formatDateTime } from "#/utils/time";
+import { AIBridgeSetupAlert } from "../AIBridgeSetupAlert";
 import { ListSessionsFilter } from "./ListSessionsFilter";
 import { ListSessionsRow } from "./ListSessionsRow";
 
@@ -70,21 +68,7 @@ export const ListSessionsPageView: FC<ListSessionsPageViewProps> = ({
 	}
 
 	if (!isAISessionsEnabled) {
-		return (
-			<Alert className="mb-12" severity="warning" prominent>
-				<AlertTitle>
-					AI Bridge is included in your license, but not set up yet.
-				</AlertTitle>
-				<AlertDescription>
-					You have access to AI Governance, but it still needs to be setup.
-					Check out the{" "}
-					<Link href={docs("/ai-coder/ai-bridge")} target="_blank">
-						AI Bridge
-					</Link>{" "}
-					documentation to get started.
-				</AlertDescription>
-			</Alert>
-		);
+		return <AIBridgeSetupAlert />;
 	}
 
 	const utcOffset = formatDateTime(new Date(), DATE_FORMAT.UTC_OFFSET);
