+++ site/src/pages/AIBridgePage/RequestLogsPage/RequestLogsPageView.tsx
@@ -1,6 +1,6 @@
 import type { ComponentProps, FC } from "react";
 import type { AIBridgeInterception } from "#/api/typesGenerated";
-import { Alert, AlertDescription, AlertTitle } from "#/components/Alert/Alert";
+import { Alert } from "#/components/Alert/Alert";
 import { Link } from "#/components/Link/Link";
 import {
 	PaginationContainer,
@@ -16,7 +16,7 @@ import {
 } from "#/components/Table/Table";
 import { TableEmpty } from "#/components/TableEmpty/TableEmpty";
 import { TableLoader } from "#/components/TableLoader/TableLoader";
-import { docs } from "#/utils/docs";
+import { AIBridgeSetupAlert } from "../AIBridgeSetupAlert";
 import { RequestLogsFilter } from "./RequestLogsFilter/RequestLogsFilter";
 import { RequestLogsRow } from "./RequestLogsRow/RequestLogsRow";
 
@@ -42,21 +42,7 @@ export const RequestLogsPageView: FC<RequestLogsPageViewProps> = ({
 	}
 
 	if (!isRequestLogsEnabled) {
-		return (
-			<Alert className="mb-12" severity="warning" prominent>
-				<AlertTitle>
-					AI Bridge is included in your license, but not set up yet.
-				</AlertTitle>
-				<AlertDescription>
-					You have access to AI Governance, but it still needs to be setup.
-					Check out the{" "}
-					<Link href={docs("/ai-coder/ai-bridge")} target="_blank">
-						AI Bridge
-					</Link>{" "}
-					documentation to get started.
-				</AlertDescription>
-			</Alert>
-		);
+		return <AIBridgeSetupAlert />;
 	}
 
 	return (
